Some more component moving
