+++ app/_components/DebugMode.tsx
@@ -1,4 +1,5 @@
 "use client";
+
 import Link from "next/link";
 import { createContext, useContext } from "react";
